t: add flux-accounting DB v0.39.0 to testsuite
Problem: A new column is added to the "jobs" table, but the test file
that tests updating previous versions of flux-accounting DBs does not
contain the schema version of the DB before this addition.

Add the flux-accounting DB from v0.39.0 to the set of DBs tested in an
update in t1017-update-db.t.