formatter: add new JobsFormatter class
Problem: The formatter module in flux-accounting does not have a class
for formatting the output of job records which would be useful because
the flux-accounting Python bindings manually define the formatting of
these job records within its bindings.

Create a new class called JobsFormatter, a subclass of
flux.util.OutputFormat, which can take a Python format string to allow
a user to customize their output, similar to flux-jobs.