view-job-records: use new JobsFormatter class
Problem: The convert_to_str() function defines its own formatting for
job records to be returned with the view-job-records command, but the
formatter module in flux-accounting now has its own JobsFormatter class
that does this (and also includes custom formatting).

Edit the convert_to_str() function to create a JobsFormatter object,
initialized with a Python format string. Add a "format" optional
argument to the view-job-records command to allow the user to specify
a custom format string for their output.