doc: add fair-share documentation
Problem: flux-accounting doesn't really have any in-depth documentation
about how fair-share is calculated for a hierarchy of banks and users.

Add some documentation about how fair-share is calculated for a
hierarchy of banks and users in flux-accounting. Create an example
hierarchy and walk through the process of calculating fair-share for an
association.