plugin: add unpack of queue priorities
Problem: The priority plugin cannot unpack any defined integer
priorities for queues in a configuration file, but there is at least
one use case to be able to set these priorities in a TOML config file.

Add an unpack of a "accounting.queue-priorities" table in the callback
for conf.update. The table will have any queues and their associated
integer priorities. If the queue already exists in the plugin's
internal queues map, update the associated priority of the queue. If
it does not exist, add the queue to the plugin's internal map.