t: add tests for configuring queue priorities
Problem: The test suite does not contain any tests for configuring
queues in a TOML config file and ensuring a job's priority is affected
by the new configuration.

Add some tests that go through changing the priority for a queue and
ensuring that a job's priority submitted to that queue will change
depending on the configuration.