doc: update man page for flux-config-accounting
Problem: The man page for flux-config-accounting is out-of-date and
needs to be updated to include how to define priorities for queues to
be read and used by the priority plugin.

Add a section to the man page explaining how to set integer priorities
for queues in a TOML configuration file.