doc: add example on priority configs for queues
Problem: The accounting guide makes mention of setting priorities for a
queue so that it can affect the priority calculation for a submitted
job, but there are no examples this configuration anywhere in the docs.

Add a section in the Multi-Factor Priority Plugin section on how
configuring priorities for certain queues affects a job's calculated
priority. Write a short example on how the priority for a job changes
depending on this configuration.