t: add comments for what is happening in tests
Problem: Some of the tests in the testsuite can be somewhat complex and
simulate a scenario that might take multiple steps; it would be helpful
if descriptions were added that went into more detail than what is
described in the one-liner for each test.

Add some more thorough test descriptions in a couple of the test files
explaining what is being tested, particularly for the scenarios that
are broken up over multiple tests.