t1005: update test description
Problem: There is a test in t1005-max-jobs-limits.t that submits two
jobs to the priority plugin under the same bank using two different
methods: by default and by explicitly setting it with --setattr.
However, the test description is confusing and could be worded better.

Update the test description to more concisely describe what is
happening in the test.