t1011: update comment above test
Problem: One of the tests in t1011-job-archive-interface.t has a
lengthy comment explaining in more detail what is being checked in the
test itself, but it is missing an important "not" when describing the
effect on the job usage value for an association.

Fix this comment by adding a "not" to the sentence.