doc: add section on configuring queue permissions
There are no examples or explanations of how flux-accounting's priority
plugin can enforce permissions for associations on certain queues by
setting them up in flux-accounting.

Add some documentation explaining how to set this up. Include a simple
example of configuring an association's list of permissible queues and
how a job gets rejected if they submit a job to a queue they do not
have access to.