bank_table: add "max_preempt_after" column
Problem: The bank_table does not have a way to track a max
preemptible-after value for a certain bank. This kind of information is
needed to enforce a limit on association-set preemptible-after values on
their jobs as well as set a default on their job in the case where the
association does not set one on their own.

Add a new column to the bank_table called "max_preempt_after", which
is a floating point number representing the max number of seconds in
which a job submitted under this bank can run before being preempted.
It does not require a value and does not have a default.

Update the schema version number of the flux-accounting database to
account for the change to bank_table.

Add "max_preempt_after" to the list of column names defined in the
BANK_TABLE constant in __init__.py.