commands: add --max-preempt-after optional arg
Problem: The add-bank and edit-bank commands have no way to specify a
max-preempt-after value for a bank when it is being added to the
bank_table.

Add a new optional argument to the add-bank and edit-bank commands:
--max-preempt-after, which takes a Flux Standard Duration (FSD) value
representing the max amount of time in which a job submitted under this
bank can run before being preempted.

The max-preempt-after attribute for a bank can also be cleared by
passing -1 to the --max-preempt-after optional argument in the
edit-bank command.