edit-bank: add --max-preempt-after optional arg
Problem: The edit-bank command has no way to specify a
max-preempt-after value for a bank when it is being edited.

Add a new optional argument to the edit-bank command:
--max-preempt-after, which takes a Flux Standard Duration (FSD) value
representing the max amount of time in which a job submitted under this
bank can run before being preempted.

The max-preempt-after attribute for a bank can also be cleared by
passing -1 to the --max-preempt-after optional argument in the
edit-bank command.