plugin: add max_cores attribute per association
Problem: The priority plugin does not store max_cores information
per-association in the plugin, which it will need in order to do
resource tracking across all of an association's running jobs.

Add max_cores to the list of information sent to the priority plugin
per-association.

Add max_cores as an attribute to the Association class so that it can
be stored and accessed per-association in the plugin. Set a default for
this attribute when creating a temporary association in the plugin.

Edit the accounting.cpp unit test to account for the addition of the
new max_cores attribute.