plugin: copy over jj.* code
Problem: flux-core has some useful code for counting the resources for
a job, but it needs to be manually copied over.

Copy the jj.* files from flux-core and place them in the same folder as
the priority plugin so that it can be compiled together and have its
methods able to be accessed by the plugin.

Manually copy over the streq() definition and place it in jj.hpp as
well as copy over its LICENSE from flux-core. Reference it in jj.hpp.