association_table: add max_cores column
Problem: The association_table has no way to track the max number of
cores an association can have across all of their running jobs. This is
needed in order to track the max number of resources an association has
across all of their running jobs.

Add a new column to the association_table called "max_cores".

Update the schema version number for the flux-accounting database to
account for the new column addition.