plugin: add/rm dependencies due to resource usage
Problem: The priority plugin tracks resource usage across an
association's set of running jobs, but it does not enforce any limits
on this resource usage in the event where an association goes over
their limit.

Add a check in job.state.depend to see if the job's resources would put
the association over *either* their max_nodes or max_cores limit. If
so, add a dependency on the job and hold it.

In job.state.inactive, rework the check to see if there are any held
jobs for the association to also ensure the held job would satisfy
all three conditions:

1) the association is under their max-running-jobs limit
2) the job would not put the association over their max nodes limit
3) the job would not put the association over their max cores limit

before fully releasing the job to be scheduled to run. Keep track of
which held jobs have which dependencies in a map within the plugin,
where the key of the map is the ID of a held job and the value is a
Job object that stores the node/core counts for the job as well as any
dependencies associated with the job.

The dependencies added to the held job are removed as the above
conditions are satisfied. Once all dependencies are removed from the
job, it can be released by the plugin to be scheduled to run, and the
key-value pair is removed from the internal map that keeps track of
held jobs.