t: add max_nodes, max_cores fields to payload data
Problem: There are sharness tests that manually create JSON payloads to
be sent to the plugin, but these payloads don't contain max_cores or
max_nodes information per-association.

Add this information to the payloads sent to the plugin since max_cores
and max_nodes are now enforced across an association's set of running
jobs.