plugin: add new Job class
Problem: The priority plugin needs a way to keep track of held jobs due
to either max-running-jobs or max-resource limits across an
association's set of jobs.

Create a new Job class to be utilized by the priority plugin in order to
keep track of a held job's list of dependencies and its requested
resource counts. Add this class to its own source file and compile it
along with the rest of the priority plugin code.