view-job-records: accept more timestamp formats
Problem: The --after-start-time and --before-end-time optional args for
the view-job-records command only takes one format: a
seconds-since-epoch timestamp, which isn't super user friendly.

Utilize flux.util.parse_datetime() to accept multiple, more
human-readable format strings as values to the --after-start-time and
--before-end-time optional args. Add a helper function to
jobs_table_subcommands to perform the parsing of the string. If parsing
the timestamp string fails, just attempt to pass it as a
seconds-since-epoch timestamp (which will return no job records if the
timestamp is invalid).