plugin: rm exception raise when no aux item found
Problem: the multi-factor priority plugin raises an exception on a job
in job.state.priority when it cannot find the aux item containing the
association information for the job. However, when a system instance is
restarted with pending jobs that are reprioritized, the aux items are
cleared from the job. So, the job will have an exception raised on it
even if the plugin has accounting data for the association that
submitted the job.

Remove raising a job exception in job.state.priority in the case where
the job does not have an aux item for the accounting information
associated with the job. Instead, attempt to perform another lookup for
the flux-accounting information for the association that submitted the
job.