DB: add "ON CONFLICT" clause for fairshare
Problem: The "fairshare" column in the association_table should be made
an editable column by admins, but right now it doesn't have the right
infrastructure in the database to do so because there is no reset of a
value in the case where a 'NOT NULL' constraint fails, similar to the
other editable fields in the association_table.

Add a "ON CONFLICT" clause to the fairshare column in the
association_table to match the other editable fields. Update the DB
schema to reflect the update to the association_table.