add-user: add "fairshare" as an optional arg
Problem: The "fairshare" attribute for associations is not configurable
in the add-user command, meaning that every association that gets added
to the flux-accounting database has to have a fairshare value of 0.5.
There might be a case where an admin might want to add an association
with a different fairshare value, but in order to do this, they have to
open a SQLite shell, connect to the DB, and write and execute a SQL
statement.

Add "fairshare" as an optional argument to the add-user command, which
allows an admin to customize the fairshare value for an association if
they want to. By default, this value will be set to 0.5, which is also
its default value in the database.