get_default_bank(): change "return" structure
Problem: The get_default_bank() function assumes that the SQL query
executed will work every time it is run because associations are
assumed to never be deleted from the association_table. However, there
will be a case where the row can actually be deleted, so the index
access of the "result" variable should be checked before returning.

Adjust the "return" logic of the get_default_bank() function to only
"return" the default bank if a result is actually returned from the SQL
query. If there is no value returned from the query, just return None.