delete-bank: add --force option
Problem: The delete-bank command does not actually delete a row from
the bank_table but instead sets the 'active' column for a row to
0. There might be a use case where an admin would actually want to
remove a row from the table.

Add a new --force optional argument to the delete-bank command which
will actually remove the row from the bank_table.

Add a function description to delete_bank() to describe what the
function is doing and add a note about the --force option.