delete-user: add --force option
Problem: The delete-user command does not actually delete a row from
the association table but instead sets the 'active' column for a row to
0. There might be a use case where an admin would actually want to
remove a row from the table.

Add a new --force optional argument to the delete-user command which
will actually remove the row from the association_table.

Add a function description to delete_user() to describe what the
function is doing and add a note about the --force option.