python: create new "util" folder
Problem: The flux-accounting Python bindings have utility files
(formatter.py and sql_util.py) that are used across the accounting
module, but they are distinctly different from the rest of the module;
they are better classified as utility files and should be moved to
their own folder.

Move formatter.py and sql_util.py to their own "util" folder in the
Python bindings.

Adjust the Makefile in src/bindings/python/fluxacct/ to do the
packaging for the "fluxacct" Python package since it will now grab
files from both the accounting/ and util/ folders. Remove the
Makefile.am's from both the accounting/ and util/ folders. Add an
__init__.py.in file to the fluxacct/ folder.