python: create new "util" folder
Problem: The flux-accounting Python bindings have utility files
(formatter.py and sql_util.py) that are used across the accounting
module, but they are distinctly different from the rest of the module;
they are better classified as utility files and should be moved to
their own folder.

Move formatter.py and sql_util.py to their own "util" folder in the
Python bindings. Adjust the appropriate Makefiles and configure.ac to
account for the change.