flux-account-priority-update: use column names
Problem: The flux-account-priority-update script uses row indices to
build the data that is sent to the priority plugin, but this is
problematic because it relies on internal knowledge of the DB schema
and it is unreadable.

Use the conn.row_factory property to access row values by column name
instead of by index.