bindings: use column names
Problem: The Python bindings use row indices to access the results of
a SQL query in a number of functions, but this is problematic because it
relies on internal knowledge of the DB schema and it is unreadable.

Use the conn.row_factory property to access row values by column name
instead of by index.