formatter: use column names
Problem: The formatter module uses row indices to access the results of
a SQL query in the construct_parsable_hierarchy() function, but this is
problematic because it relies on internal knowledge of the DB schema and
it is unreadable.

Use the conn.row_factory property to access row values by column name
instead of by index.