update-fshare: change job usage var to double
Problem: The variable that stores job usage values for banks in the
update-fshare script is a 32-bit integer, and it cannot keep up with
larger job usage values that are generated on big machines.

Further, the job usage value for an association or bank is not always
guaranteed to be a whole number, and placing it an integer variable
makes the fair-share calculation less accurate.

Change the variable type of the job usage value from a 32-bit integer
to a 64-bit double.