.github: update Ubuntu version for actions
Problem: There are a couple of GitHub actions that use ubuntu-20.04, but
this version will no longer be supported in April 2025.

They also use versions of Python that are not found on later versions
of Ubuntu.

The pylint GitHub action also no longer works with a newer version of
Python > 3.10: https://stackoverflow.com/questions/78459119

Update the actions to use ubuntu-latest. Update the version of Python
used in the python format and pylint actions to 3.9 to ensure
compatibility with the latest Ubuntu version.