data_reader_db: change "shares" type to uint64_t
Problem: The "shares" attribute for the Account class is of type
uint64_t, but data_reader_db is extracting the shares value for an
association as a string and then casting it back to a long long. It
would make more sense to just set this variable type as a uint64_t from
the start.

Change the variable type of the shares value to uint64_t to match its
type in the Account class.