fairshare: change "usage" type to double
Problem: The "usage" attribute used throughout the fair-share code is of
type uint64_t, but it is of type "real" (i.e a floating-point value) in
the SQLite database. This means that the job usage values are getting
truncated when they are being passed through to the actual weight -->
fair-share calculation. It would be more precise to keep this "usage"
value as type double throughout the fair-share calculation code.

Change the type of "usage" throughout the fair-share code to be of type
double.