view-queue: use new QueueFormatter class
Problem: The view-queue command defines its own formatting for the
output of the command, but the formatter module now has its own
QueueFormatter subclass.

Refactor the view_queue() function to use the new QueueFormatter class.
Add a --parsable option to the view-queue command to align with the
other view-* commands in the event where a user wants to see all of the
information for a queue on one line.