cmd: add new list-queues command
Problem: There is no way to concisely list a set of defined queues in
the flux-accounting database, similar to the list-banks and
list-projects command.

Add a new command to the command suite called list-queues, which will
list all of the queues defined in the queue_table. Offer the same
options as list-banks to customize the output.