update_bank_cb: raise specific jobtap_error(s)
Problem: The error message that update_bank_cb () raises when it cannot
find flux-accounting information for a particular association is vague
and hard to debug, for there could be multiple reasons: the plugin
could still be waiting on flux-accounting data to be loaded in OR there
exists no flux-accounting data for that particular association.

Rework the handling in update_bank_cb () in the case where the plugin
cannot find information for a particular association; specifically,
check to see if the plugin has ANY flux-accounting data before
returning an error, and customize the error messages returned to be more
specific depending on the case.