t: update expected error messages
Problem: The error message in update_bank_cb () has been updated to
give slightly more context when it fails to update the bank for a job,
but there are a couple of tests in the testsuite that still check for
the old message.

Update the expected error messages in t1032 and t1033 to account for
the updated error message.