doc: add sanity checklist for flux-accounting
Problem: flux-accounting contains a guide for setting up
flux-accounting, but it does not contain any documentation for checking
that the various components of flux-accounting are up, configured, and
running correctly.

Add a "Health/Sanity Checklist" to flux-accounting that walks through
checking the various components of the accounting system and making
sure it is running as expected.