decay factor: fix iteration through usage periods
Problem: The for-loop that iterates through the decayed job usage
factors for an association starts and stops at the wrong index of the
list that is passed in. It currently starts at the *second* index and
stops at the second-to-last index, but it should be starting at the
*first* index and stop at the second-to-last index (since the last
index gets discarded when a new usage period is reached).

Change the starting index for the for-loop in apply_decay_factor() to
start at the *first* index.