apply_decay_factor: commit UPDATE statements
Problem: The UPDATE SQL statements executed in apply_decay_factor are
not actually applied when this function is called because they are not
committed by the time the function exits. Therefore, decays are not
being applied for associations as usage periods go on.

Commit the UPDATE statements that occur in apply_decay_factor() to
actually apply them to the flux-accounting database.