t: fix expected usage factor calculated for user
Problem: The usage factor calculated for a user after a half-life
period has passed is incorrect because of the way apply_decay_factor()
was applying historical job usage factors. Now that the function
correctly calculates and applies past usage factors to each
association, the expected job usage factor needs to be adjusted in the
unit tests.

Adjust the expected value in the unit tests where we are calculating a
job usage factor for an association after a half-life period has passed
(and previous job usage factors experience a decay).