jj_read_vertex (): ignore unknown resources
Problem: The jj_read_vertex () function returns an error if it
encounters a resource type it does not recognize certain types like
'ssd'. Since flux-coral2 can create non-v1 jobspecs, flux-accounting
needs to be able to handle these types of resources instead of raising
an exception on the job.

Remove the else-condition from jj_read_vertex () to raise an error when
it encounters an unknown resource type, and instead just ignore unknown
resources.