cmd: add a new "list-users" command
Problem: There is no way to list any of the rows in the
association_table, similar to the "list-banks" command. It would also
be nice if the table could be filtered by certain columns in the table,
such as by queue, project, or limit.

Add a new command to the flux-accounting command suite called
"list-users", which by default will list all of the rows in the
association_table in a human-readable table format. An optional --json
argument is also added to output the results in JSON format.

This command accepts multiple optional arguments to filter the table by
certain columns in the association_table.

Its output can also be customized by column, e.g the person calling this
command only wants to see the username of the association.