formatter: add as_format_string() method
Problem: There is no way to customize the output of a query to the
flux-accounting DB using a Python format string. It would be nice if
flux-accounting offered a way to customize the output with a format
string in addition to the --parsable and --fields options.

Add a new method to the AccountingFormatter class called
as_format_string(), which will return the results of a query to the
flux-accounting DB as a formatted string.