view-* commands: add -o/--format optional argument
Problem: The view-* commands do not offer an optional argument to view
the output as a formatted string. It would be nice if the command
offered a way for the user to customize the output of these commands.

Add an -o/--format optional argument to the view-* commands, which allow
the user to pass in a formatted string to customize their output.