list-* commands: add -o/--format optional arg
Problem: The list-* commands do not offer an optional argument to
view the output as a formatted string. It would be nice if the command
offered a way for the caller to customize the output of this command.

Add an -o/--format optional argument to the list-* commands, which
allow the user to pass in a formatted string to customize their
output.