formatter: add ProjectFormatter subclass
Problem: the view-project and list-projects commands manually define
their own formatting for their respective commands, which is different
from the rest of the command suite, which are subclasses of a common
AccountingFormatter class to standardize all output.

Add a new ProjectFormatter subclass to the Python bindings so that the
output of the view-project and list-projects commands are standard with
the rest of the flux-accounting command suite. Add --table and --fields
options to the list-projects command. Add a --parsable option to the
view-projects command.