view-project/list-projects: add -o/--format option
Problem: The view-project and list-projects commands do not take an
optional -o/--format argument to format the output using a Python
format string, which is inconsistent with the rest of the
flux-accounting command suite.

Add an -o/--format optional argument to the view-project/list-projects
commands, which allows the user to format the output of either command
using a Python format string.