job.state.depend: move comment
Problem: Most comments in the plugin that explain if-conditions are
written underneath the if-statement, but there is one in
job.state.depend that is written above the statement.

Move it to below the statement (and edit it slightly) so that it follows
the convention of the other comments in the plugin.