job.state.depend: remove return from first check
Problem: The check to see if the association already has the max number
of jobs running in a certain queue has a "return 0" statement at the end
of the check, but there is potential for there to be multiple
dependencies added to the job, e.g the association could be at the max
number of jobs running in a certain queue AND they could also be at
their more-general max running jobs limit.

Remove the "return 0" statement from the end of the first dependency
check for a max running jobs limit in a queue.