plugin: create label for dependency_remove < 0
Problem: The dependency_remove () calls in the callback for
job.state.inactive are a little tough to read in the case that an
exception needs to be raised because they are broken up over multiple
lines. As more dependencies are added to this callback, this function
will get even longer and perhaps more tedious to read through.

Add an "error" label to the job.state.inactive callback that is called
when removing a dependency fails. Move the
flux_jobtap_raise_exception () call to this label and set the name of
the dependencies before going to this label to include it in the
exception message.