accounting: move Queue class definition to top
Problem: The Queue class definition is below the Association class
definition, but there will eventually be a need to pass a Queue object
to a method in the Association class. There is also no default
initialization for the attributes of a Queue object.

Move the macro definitions to the top of accounting.hpp. Move the Queue
class definition to the top of accounting.hpp. Add some default values
to the various attributes of a Queue object; for the min_nodes_per_job
limit and priority, set it to 0 if no value is specified when the
object is created. For the max_* limits, set it to a very large integer
if no value is specified.