Queue: add "name" attribute
Problem: The Queue class does not have a "name" attribute which holds
the name of the queue. This would be nice to have for future Queue and
Association methods.

Add a "name" attribute to the Queue class. Adjust the callback that
unpacks queue information from the flux-accounting DB as well as the
Accounting and Queue object unit tests to also store the name of the
queue when creating Queue objects.