test: remove plan () call for accounting unit test
Problem: The main () function for the unit tests for the Association
and Queue classes has a plan () call at the beginning to explicitly lay
out how many unit tests it wants to run, but as more and more unit
tests are added, this function call has to change to reflect the new
number of tests and is tedious to keep track of.

Remove the plan () call from main ().