plugin: don't use max_run_jobs_per_queue ()
Problem: The plugin currently calls a max_run_jobs_per_queue ()
function to find the max_running_jobs attribute for a Queue object, but
now the Queue class has an appropriate default (a very large integer)
when a Queue object is initialized as well as a name attribute. So, the
lookup for this attribute can be made directly in the plugin with a
one-liner map key-value access.

Remove the max_run_jobs_per_queue () function from accounting.hpp and
accounting.cpp in favor of just using the one-liner in the plugin.