to_json (): restructure held_jobs object
Problem: the held_jobs attribute of an Association object has changed
from a vector of job IDs to a vector of Job objects. The to_json ()
method has to change to include all of the attributes of a Job object in
the output for each association.

Change the to_json () method to account for the change to held_jobs;
specifically, add all of the various attributes for a Job object to each
association's JSON object when constructing the output for the
jobtap.query callback.