delete-queue: issue a WARNING when deleting queue
Problem: A queue can be deleted from the queue_table but still be
referenced by associations in the association_table, i.e. an
association can still list that queue in its "queues" column. An admin
who calls the delete-queue command should be warned that they should
also edit any association rows in the association_table to account for
deleting a queue.

Check to see if any associations have this queue listed in their
"queues" column before deleting the queue. If so, issue a WARNING
statement that the admin should also edit any association's rows to
account for the deletion.