to_json (): add error label
Problem: There are many error checks in the to_json () method that
json_decref () previously initialized json_t * objects. This produces a
lot of duplicated code as the function progresses because as more json_t
* objects get created, more have to be decref'd. It would be cleaner to
just have an error label.

Add an error label to the to_json () method that will call json_decref
() on all of the json_t * objects created in the method.