plugin: abstract checks for dependencies
Problem: The priority plugin has to manually extract and check info
about both the job and the association that is submitting the job for
any potential flux-accounting limits in both job.state.depend and
job.state.inactive. As more and more dependencies are added to keep
track of in flux-accounting, it will get more cumbersome to have to do
these checks in the callbacks themselves.

Utilize the new Job class and the methods provided in the Association
class to do comparisons of an association's flux-accounting limits when
their job enters job.state.depend and job.state.inactive.

Change the held_jobs member from a vector of just Flux job IDs to a
vector of class Job, which will store attributes like its job ID, an
optional queue, and any resources counted for the job. As a result,
remove the queue_held_jobs member from the Association class since it
is no longer needed.

In job.state.depend, check each dependency one at a time and push any
and all dependencies back onto the Job object's dependencies vector. If
any dependencies exist on the job, store the Job object in the
Association object.

In job.state.inactive, when an association's currently running job
completes, check their held_jobs vector for any held jobs and check to
see if any jobs now satisfy all requirements to be released from any
flux-accounting dependencies.