Association: add methods for limit checks, macros
Problem: The callbacks for job.state.depend and job.state.inactive
manually check the limits for an association when looking to see if a
held job should be released. As more and more flux-accounting limits
begin to be enforced in the plugin, it will get more difficult to
efficiently add these checks in the callback itself. It would be cleaner
to abstract these checks out of the plugin code, where they could be
modularized and unit tested.

Add two new methods to the Association class that check an association's
limits that are currently enforced in the plugin: a max-running-jobs
limit per-queue and a max-running-jobs limit per-association.

Create macro definitions for the flux-accounting dependency names used
in the plugin and store them in the accounting header file.