plugin: abstract loop for association's held jobs
Problem: The job.state.inactive callback has a loop defined in the
callback itself to check an association's list of held jobs to see if
any and all of them meet all flux-accounting limits in order to be
released, but this loop is complex and makes reading the callback
difficult. It would be cleaner if this loop was abstracted to a helper
function.

Move the while-loop to iterate through an association's list of held
jobs to see if any of them can be released by the plugin to its own
function. Adjust the error label in job.state.inactive to be more
general since the error label in the helper function will have a more
specific error message as to what dependency failed to be removed from
the job.