construct_hierarchy (): use column names
Problem: The construct_hierarchy () method uses column indices instead
of column names when placing the results of a SQL query into the string
constructed for a flux-accounting bank/user hierarchy. This makes the
code hard to parse and understand exactly what fields are being placed
where in the string.

Change the column indices to instead use column names when accessing
the results of a SQL query and formatting the resulting string.